requirements.txt
torch
torchvision
pytorch-metric-learning
numpy
pandas
Pillow
imageio
